pyproject.toml
[project]
name = "spark-your-spotify"
version = "0.1.0"
description = "CLI tool that discovers local artists based on your Spotify taste"
requires-python = ">=3.9"
dependencies = [
    "spotipy>=2.23.0",
    "requests>=2.31.0",
    "jinja2>=3.1.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
]

[project.scripts]
spark = "src.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
